# api/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_dso_preview_api_transformers_dso_csv_preview_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_dso_preview_api_transformers_dso_csv_preview_post",
        "type": "object"
      },
      "BulkUpdate": {
        "properties": {
          "capacity_distribution": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Distribution"
          },
          "confirm": {
            "default": false,
            "title": "Confirm",
            "type": "boolean"
          },
          "method": {
            "pattern": "^(uniform|percentage)$",
            "title": "Method",
            "type": "string"
          },
          "plz": {
            "title": "Plz",
            "type": "integer"
          },
          "transformer_rated_power": {
            "anyOf": [
              {
                "maximum": 100000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transformer Rated Power"
          }
        },
        "required": [
          "plz",
          "method"
        ],
        "title": "BulkUpdate",
        "type": "object"
      },
      "CapacityUpdate": {
        "properties": {
          "transformer_rated_power": {
            "maximum": 100000.0,
            "minimum": 1.0,
            "title": "Transformer Rated Power",
            "type": "integer"
          }
        },
        "required": [
          "transformer_rated_power"
        ],
        "title": "CapacityUpdate",
        "type": "object"
      },
...
      "ClearCapacities": {
        "properties": {
          "confirm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirm"
          },
          "plz": {
            "title": "Plz",
            "type": "integer"
          }
        },
        "required": [
          "plz"
        ],
        "title": "ClearCapacities",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewTransformer": {
        "properties": {
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "plz": {
            "title": "Plz",
            "type": "integer"
          },
          "transformer_rated_power": {
            "anyOf": [
              {
                "maximum": 100000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transformer Rated Power"
          }
        },
        "required": [
          "plz",
          "lon",
          "lat"
        ],
        "title": "NewTransformer",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API of the pylovo workflow (database, regions, data, generation, results) for the GridPlanner UI.",
    "title": "pylovo-api",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/transformers": {
      "get": {
        "description": "Transformers of a PLZ (spatial intersection with its polygon) or of a map bounding box.",
        "operationId": "list_transformers_api_transformers_get",
        "parameters": [
          {
            "in": "query",
            "name": "plz",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plz"
            }
          },
          {
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bbox"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Transformers Api Transformers Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Transformers",
        "tags": [
          "transformers"
        ]
      },
      "post": {
        "description": "Add a manual transformer position (``osm_id = manual/<epoch-ms>``).",
        "operationId": "add_transformer_api_transformers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTransformer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Transformer Api Transformers Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Transformer",
        "tags": [
          "transformers"
        ]
      }
    },
    "/api/transformers/bulk": {
      "post": {
        "description": "Set all capacities of a PLZ to one size, or distribute sizes randomly by percentage.",
        "operationId": "bulk_update_api_transformers_bulk_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Bulk Update Api Transformers Bulk Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Bulk Update",
        "tags": [
          "transformers"
        ]
      }
    },
...
    "/api/transformers/clear-capacities": {
      "post": {
        "description": "Remove the rated power of every transformer in a PLZ.",
        "operationId": "clear_capacities_api_transformers_clear_capacities_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClearCapacities"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Clear Capacities Api Transformers Clear Capacities Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Clear Capacities",
        "tags": [
          "transformers"
        ]
      }
    },
    "/api/transformers/dso-csv/preview": {
      "post": {
        "description": "Store an uploaded DSO CSV and return a validation preview (nothing is imported yet).",
        "operationId": "dso_preview_api_transformers_dso_csv_preview_post",
        "parameters": [
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_dso_preview_api_transformers_dso_csv_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dso Preview Api Transformers Dso Csv Preview Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dso Preview",
        "tags": [
          "transformers"
        ]
      }
    },
    "/api/transformers/osm-relations": {
      "get": {
        "description": "Find OSM boundary relations by name (Nominatim) to fill in ``--relation-id``.",
        "operationId": "osm_relations_api_transformers_osm_relations_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 120,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Osm Relations Api Transformers Osm Relations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Osm Relations",
        "tags": [
          "transformers"
        ]
      }
    },
    "/api/transformers/{osm_id}": {
      "delete": {
        "description": "Delete one transformer. Positions of generated grids that use it are removed as well.",
        "operationId": "delete_transformer_api_transformers__osm_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "osm_id",
            "required": true,
            "schema": {
              "title": "Osm Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Transformer Api Transformers  Osm Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Transformer",
        "tags": [
          "transformers"
        ]
      },
      "patch": {
        "description": "Set the rated power of one transformer.",
        "operationId": "update_capacity_api_transformers__osm_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "osm_id",
            "required": true,
            "schema": {
              "title": "Osm Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapacityUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Capacity Api Transformers  Osm Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Capacity",
        "tags": [
          "transformers"
        ]
      }
    },
...
  }
}